% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readbed}
\alias{readbed}
\title{Read BED Files}
\usage{
readbed(bed, obs, snp)
}
\arguments{
\item{bed}{Path to the BED file.}

\item{obs}{Total number of subjects (rows) in the FAM file.}

\item{snp}{1-based index of the SNP to read.}
}
\value{
Numeric vector of genotypes (0, 1, 2, or NA) at the selected SNP.
}
\description{
Low-level reader for a single SNP from a PLINK binary BED file (SNP-major
mode).
}
