% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageHelp.R
\docType{package}
\name{ReadPlink-package}
\alias{ReadPlink}
\alias{ReadPlink-package}
\title{ReadPlink: Read PLINK binary genotype files}
\description{
Read compressed PLINK binary genotype files (BED/BIM/FAM) into R.
Provides \code{\link{ReadGeno}} for reading genotypes into a matrix,
\code{\link{ReadPlinkData}} for genotypes plus BIM/FAM metadata,
\code{\link{readbed}} for low-level single-SNP access, and
\code{\link{SamplePath}} for the path to bundled sample data.
}
\author{
Zachary R. McCaw
}
\keyword{internal}
