% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageHelp.R
\docType{package}
\name{ReadPlink}
\alias{ReadPlink}
\alias{ReadPlink-package}
\title{ReadPlink: Reads Plink Data into R}
\description{
Reads compressed plink data into R in additive format. See \code{\link{ReadGeno}}.
}
\author{
Zachary R. McCaw
}
