% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readbed}
\alias{readbed}
\title{Read BED Files}
\usage{
readbed(bed, obs, snp)
}
\arguments{
\item{bed}{Name of the BED file.}

\item{obs}{Total number of subjects in FAM file.}

\item{snp}{SNP to import.}
}
\value{
Numeric vector of genotypes at the selected SNP.
}
\description{
Read BED Files
}
