% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readbed}
\alias{readbed}
\title{Read BED Files}
\usage{
readbed(bed, obs, snp)
}
\arguments{
\item{bed}{Name of the bed file.}

\item{obs}{Number of subjects.}

\item{snp}{Locus to select.}
}
\value{
Numeric vector of genotypes at the select locus.
}
\description{
Read BED Files
}
