% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGeno.R
\name{ReadGeno}
\alias{ReadGeno}
\title{Read Genotypes from Plink}
\usage{
ReadGeno(stem, loci = NULL, subj = NULL, keep = F, parallel = F)
}
\arguments{
\item{stem}{Stem of the BED, BIM, FAM files.}

\item{loci}{Row numbers of desired loci in the BIM file. If omitted, all loci
are returned.}

\item{subj}{Row numbers of desired subjects in the FAM files. If omitted, all
subjects are turned.}

\item{keep}{Return BIM and FAM files in addition to genotypes?}

\item{parallel}{Run in parallel? Must register parallel backend first.}
}
\value{
Either a numeric genotype matrix, or a list containing the genotype
  matrix together with the imported BIM and FAM files.
}
\description{
Reads compressed plink genotypes into R. Specify the \code{stem} of the input
files, excluding extensions. The user can choose to retain only certain loci 
from the BIM file, or subjects from the FAM file. Genotypes are formatted
into a numeric matrix, with subjects as rows and loci as columns. If \code{keep=T},
the BIM and FAM files for selected loci and subjects are additionally returned. 
If \code{parallel=T}, the genotypes are imported in parallel.
}
